requirements.txt
async-generator==1.10
attrs==22.1.0
certifi==2021.10.8
cffi==1.15.1
charset-normalizer==2.1.1
greenlet==1.1.3
h11==0.13.0
helium==3.0.8
idna==3.3
lxml==4.9.1
numpy==1.23.2
outcome==1.2.0
pandas==1.4.4
psycopg2-binary==2.9.5
pycparser==2.21
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==0.21.0
pytz==2022.2.1
selenium==3.141.0
six==1.16.0
sniffio==1.2.0
sortedcontainers==2.4.0
SQLAlchemy==1.4.41
trio==0.21.0
trio-websocket==0.9.2
urllib3==1.26.11
wsproto==1.1.0
flask== 2.3.3
Flask-SQLAlchemy==3.0.5